# api/openapi.yaml
openapi: 3.0.0
info:
  title: Calculator
  description: This is a calcultor API
  contact:
    email: omer72@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1"
servers:
  - url: https://virtserver.swaggerhub.com/OmerEtrog/Calculator/1
    description: SwaggerHub API Auto Mocking
  - url: http://localhost:8080
    description: SwaggerHub API Auto Mocking
tags:
  - name: calculate
    description: Simple Calculator
  - name: generateToken
    description: Generate Token
paths:
  /generateToken:
    get:
      tags:
        - generateToken
      summary: get the token
      description: Get a token
      operationId: generateToken
      responses:
        "201":
          description: Successful operation
        "400":
          description: Failed to get the token
      x-swagger-router-controller: GenerateToken
  /calculate:
    post:
      tags:
        - calculate
      summary: calculate 2 numbers using the action
      description: Calculate 2 numbers using the action
      operationId: calculate
      requestBody:
        description: num1, num2 and Operation
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculatorValues'
        required: true
      responses:
        "201":
          description: Successful operation
        "400":
          description: "invalid input, object invalid"
        "401":
          description: Access token is missing or invalid
      security:
        - bearerAuth: []
      x-swagger-router-controller: Calculator
components:
  schemas:
    CalculatorValues:
      required:
        - num1
        - num2
        - operation
      type: object
      properties:
        num1:
          type: number
        num2:
          type: number
        operation:
          type: string
          enum:
            - add
            - subtract
            - multiply
            - divide
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT